event_consumer/requirements.txt
fastapi
sqlalchemy
pydantic
python-multipart
aiofiles
python-dotenv
python-jose
bcrypt
passlib
aio-pika
pika
uvicorn
argon2_cffi
alembic
websockets
psycopg2-binary
requests
minio
sentry-sdk

pymongo